+++ api-types/src/vm_config/memory_config.rs
@@ -1,0 +1,32 @@
+// Copyright © 2026 Contributors to the Cloud Hypervisor project
+//
+// SPDX-License-Identifier: Apache-2.0
+//
+
+use std::str::FromStr;
+
+use serde::{Deserialize, Serialize};
+
+#[derive(Clone, Copy, Debug, PartialEq, Eq, Deserialize, Serialize, Default)]
+pub enum HotplugMethod {
+    #[default]
+    Acpi,
+    VirtioMem,
+}
+
+#[derive(Debug)]
+pub enum ParseHotplugMethodError {
+    InvalidValue(String),
+}
+
+impl FromStr for HotplugMethod {
+    type Err = ParseHotplugMethodError;
+
+    fn from_str(s: &str) -> Result<Self, Self::Err> {
+        match s.to_lowercase().as_str() {
+            "acpi" => Ok(HotplugMethod::Acpi),
+            "virtio-mem" => Ok(HotplugMethod::VirtioMem),
+            _ => Err(ParseHotplugMethodError::InvalidValue(s.to_owned())),
+        }
+    }
+}
