+++ vmm/src/memory_manager.rs
@@ -21,7 +21,7 @@ use std::{cmp, ffi, panic, result, thread, time};
 
 use acpi_tables::{Aml, aml};
 use anyhow::{Context, anyhow};
-use api_types::MemoryRestoreMode;
+use api_types::{HotplugMethod, MemoryRestoreMode};
 use arch::{RegionType, layout};
 #[cfg(target_arch = "x86_64")]
 use devices::ioapic;
@@ -62,7 +62,7 @@ use crate::uffd::{
     self, FaultResolution, FileUffdMemorySource, SocketUffdMemorySource, UffdMemorySource,
     UffdRange,
 };
-use crate::vm_config::{HotplugMethod, MemoryConfig, MemoryZoneConfig};
+use crate::vm_config::{MemoryConfig, MemoryZoneConfig};
 use crate::{GuestMemoryMmap, GuestRegionMmap, MEMORY_MANAGER_SNAPSHOT_ID, userfaultfd};
 
 struct UffdHandler {
